LPS-198943 Merge SkuHelper with SkuResourceImpl
