--- modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/helper/v1_0/SkuHelper.java
@@ -1,113 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.headless.commerce.admin.catalog.internal.helper.v1_0;
-
-import com.liferay.commerce.product.model.CPDefinition;
-import com.liferay.commerce.product.model.CPInstance;
-import com.liferay.commerce.product.service.CPDefinitionService;
-import com.liferay.commerce.product.service.CPInstanceService;
-import com.liferay.headless.commerce.admin.catalog.dto.v1_0.Sku;
-import com.liferay.headless.commerce.admin.catalog.internal.dto.v1_0.converter.constants.DTOConverterConstants;
-import com.liferay.petra.function.UnsafeConsumer;
-import com.liferay.petra.function.UnsafeFunction;
-import com.liferay.portal.kernel.search.Document;
-import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.SearchContext;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.search.filter.Filter;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
-import com.liferay.portal.vulcan.dto.converter.DTOConverter;
-import com.liferay.portal.vulcan.dto.converter.DefaultDTOConverterContext;
-import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.pagination.Pagination;
-import com.liferay.portal.vulcan.util.SearchUtil;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Locale;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Riccardo Ferrari
- */
-@Component(service = SkuHelper.class)
-public class SkuHelper {
-
-	public Page<Sku> getSkusPage(long id, Locale locale, Pagination pagination)
-		throws Exception {
-
-		CPDefinition cpDefinition =
-			_cpDefinitionService.fetchCPDefinitionByCProductId(id);
-
-		if (cpDefinition == null) {
-			return Page.of(Collections.emptyList());
-		}
-
-		List<CPInstance> cpInstances =
-			_cpInstanceService.getCPDefinitionInstances(
-				cpDefinition.getCPDefinitionId(), WorkflowConstants.STATUS_ANY,
-				pagination.getStartPosition(), pagination.getEndPosition(),
-				null);
-
-		int totalItems = _cpInstanceService.getCPDefinitionInstancesCount(
-			cpDefinition.getCPDefinitionId(), WorkflowConstants.STATUS_ANY);
-
-		return Page.of(toSKUs(cpInstances, locale), pagination, totalItems);
-	}
-
-	public Page<Sku> getSkusPage(
-			long companyId, String search, Filter filter, Pagination pagination,
-			Sort[] sorts,
-			UnsafeFunction<Document, Sku, Exception> transformUnsafeFunction)
-		throws Exception {
-
-		return SearchUtil.search(
-			null, booleanQuery -> booleanQuery.getPreBooleanFilter(), filter,
-			CPInstance.class.getName(), search, pagination,
-			queryConfig -> queryConfig.setSelectedFieldNames(
-				Field.ENTRY_CLASS_PK),
-			new UnsafeConsumer() {
-
-				public void accept(Object object) throws Exception {
-					SearchContext searchContext = (SearchContext)object;
-
-					searchContext.setAttribute(
-						Field.STATUS, WorkflowConstants.STATUS_ANY);
-					searchContext.setCompanyId(companyId);
-				}
-
-			},
-			sorts, transformUnsafeFunction);
-	}
-
-	public List<Sku> toSKUs(List<CPInstance> cpInstances, Locale locale)
-		throws Exception {
-
-		List<Sku> skus = new ArrayList<>();
-
-		for (CPInstance cpInstance : cpInstances) {
-			skus.add(
-				_skuDTOConverter.toDTO(
-					new DefaultDTOConverterContext(
-						cpInstance.getCPInstanceId(), locale)));
-		}
-
-		return skus;
-	}
-
-	@Reference
-	private CPDefinitionService _cpDefinitionService;
-
-	@Reference
-	private CPInstanceService _cpInstanceService;
-
-	@Reference(target = DTOConverterConstants.SKU_DTO_CONVERTER)
-	private DTOConverter<CPInstance, Sku> _skuDTOConverter;
-
-}
