+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-catalog-impl/src/main/java/com/liferay/headless/commerce/admin/catalog/internal/resource/v1_0/SkuResourceImpl.java
@@ -28,7 +28,6 @@ import com.liferay.headless.commerce.admin.catalog.dto.v1_0.SkuSubscriptionConfi
 import com.liferay.headless.commerce.admin.catalog.dto.v1_0.SkuVirtualSettings;
 import com.liferay.headless.commerce.admin.catalog.internal.dto.v1_0.converter.constants.DTOConverterConstants;
 import com.liferay.headless.commerce.admin.catalog.internal.dto.v1_0.util.CustomFieldsUtil;
-import com.liferay.headless.commerce.admin.catalog.internal.helper.v1_0.SkuHelper;
 import com.liferay.headless.commerce.admin.catalog.internal.odata.entity.v1_0.SkuEntityModel;
 import com.liferay.headless.commerce.admin.catalog.internal.util.DateConfigUtil;
 import com.liferay.headless.commerce.admin.catalog.internal.util.v1_0.SkuUtil;
@@ -36,10 +35,14 @@ import com.liferay.headless.commerce.admin.catalog.internal.util.v1_0.SkuVirtual
 import com.liferay.headless.commerce.admin.catalog.resource.v1_0.SkuResource;
 import com.liferay.headless.commerce.core.util.DateConfig;
 import com.liferay.headless.commerce.core.util.ServiceContextHelper;
+import com.liferay.petra.function.UnsafeConsumer;
+import com.liferay.petra.function.UnsafeFunction;
 import com.liferay.portal.configuration.module.configuration.ConfigurationProvider;
 import com.liferay.portal.kernel.change.tracking.CTAware;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
+import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.search.filter.Filter;
 import com.liferay.portal.kernel.service.ServiceContext;
@@ -65,7 +68,9 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 
 import javax.ws.rs.core.MultivaluedMap;
@@ -138,7 +143,7 @@ public class SkuResourceImpl extends BaseSkuResourceImpl {
 					externalReferenceCode);
 		}
 
-		return _skuHelper.getSkusPage(
+		return _getSkusPage(
 			cpDefinition.getCProductId(),
 			contextAcceptLanguage.getPreferredLocale(), pagination);
 	}
@@ -149,7 +154,7 @@ public class SkuResourceImpl extends BaseSkuResourceImpl {
 			@NestedFieldId(value = "productId") Long id, Pagination pagination)
 		throws Exception {
 
-		return _skuHelper.getSkusPage(
+		return _getSkusPage(
 			id, contextAcceptLanguage.getPreferredLocale(), pagination);
 	}
 
@@ -179,7 +184,7 @@ public class SkuResourceImpl extends BaseSkuResourceImpl {
 			String search, Filter filter, Pagination pagination, Sort[] sorts)
 		throws Exception {
 
-		return _skuHelper.getSkusPage(
+		return _getSkusPage(
 			contextCompany.getCompanyId(), search, filter, pagination, sorts,
 			document -> _toSku(
 				GetterUtil.getLong(document.get(Field.ENTRY_CLASS_PK)), null));
@@ -313,6 +318,54 @@ public class SkuResourceImpl extends BaseSkuResourceImpl {
 			sku.getCustomFields(), contextAcceptLanguage.getPreferredLocale());
 	}
 
+	private Page<Sku> _getSkusPage(
+			long id, Locale locale, Pagination pagination)
+		throws Exception {
+
+		CPDefinition cpDefinition =
+			_cpDefinitionService.fetchCPDefinitionByCProductId(id);
+
+		if (cpDefinition == null) {
+			return Page.of(Collections.emptyList());
+		}
+
+		List<CPInstance> cpInstances =
+			_cpInstanceService.getCPDefinitionInstances(
+				cpDefinition.getCPDefinitionId(), WorkflowConstants.STATUS_ANY,
+				pagination.getStartPosition(), pagination.getEndPosition(),
+				null);
+
+		int totalItems = _cpInstanceService.getCPDefinitionInstancesCount(
+			cpDefinition.getCPDefinitionId(), WorkflowConstants.STATUS_ANY);
+
+		return Page.of(_toSKUs(cpInstances, locale), pagination, totalItems);
+	}
+
+	private Page<Sku> _getSkusPage(
+			long companyId, String search, Filter filter, Pagination pagination,
+			Sort[] sorts,
+			UnsafeFunction<Document, Sku, Exception> transformUnsafeFunction)
+		throws Exception {
+
+		return SearchUtil.search(
+			null, booleanQuery -> booleanQuery.getPreBooleanFilter(), filter,
+			CPInstance.class.getName(), search, pagination,
+			queryConfig -> queryConfig.setSelectedFieldNames(
+				Field.ENTRY_CLASS_PK),
+			new UnsafeConsumer() {
+
+				public void accept(Object object) throws Exception {
+					SearchContext searchContext = (SearchContext)object;
+
+					searchContext.setAttribute(
+						Field.STATUS, WorkflowConstants.STATUS_ANY);
+					searchContext.setCompanyId(companyId);
+				}
+
+			},
+			sorts, transformUnsafeFunction);
+	}
+
 	private Sku _toSku(
 			Long cpInstanceId, CPInstanceUnitOfMeasure cpInstanceUnitOfMeasure)
 		throws Exception {
@@ -328,6 +381,21 @@ public class SkuResourceImpl extends BaseSkuResourceImpl {
 			"cpInstanceUnitOfMeasure", cpInstanceUnitOfMeasure);
 
 		return _skuDTOConverter.toDTO(defaultDTOConverterContext);
+	}
+
+	private List<Sku> _toSKUs(List<CPInstance> cpInstances, Locale locale)
+		throws Exception {
+
+		List<Sku> skus = new ArrayList<>();
+
+		for (CPInstance cpInstance : cpInstances) {
+			skus.add(
+				_skuDTOConverter.toDTO(
+					new DefaultDTOConverterContext(
+						cpInstance.getCPInstanceId(), locale)));
+		}
+
+		return skus;
 	}
 
 	private Page<Sku> _toUnitOfMeasureSkusPage(Page<Long> cpInstanceIdsPage)
@@ -660,9 +728,6 @@ public class SkuResourceImpl extends BaseSkuResourceImpl {
 
 	@Reference(target = DTOConverterConstants.SKU_DTO_CONVERTER)
 	private DTOConverter<CPInstance, Sku> _skuDTOConverter;
-
-	@Reference
-	private SkuHelper _skuHelper;
 
 	@Reference
 	private UniqueFileNameProvider _uniqueFileNameProvider;
